Use position fixed to put everything together in place of flex
