+++ src/narrow-screen/components/MenuListNav.js
@@ -1,5 +1,5 @@
 import React, { useContext } from 'react';
-import RestaurantContext from '../restaurant-context';
+import RestaurantContext from '../../restaurant-context';
 import styled from 'styled-components';
 
 const SideNav = styled.div`
@@ -7,9 +7,9 @@ const SideNav = styled.div`
   background: #F2F3F5;
   transform: ${props => props.open ? 'translateX(0)' : 'translateX(-100%)'};
   height: 100%;
-  z-index: 16;
+  z-index: 150;
   text-align: left;
-  position: absolute;
+  position: fixed;
   top: 0;
   left: 0;
   transition: transform 0.3s ease-in-out;
